push

+++ demo/src/main/java/com/hoanghaidang/social_network/service/impl/LikeService.java
@@ -12,29 +12,47 @@ import lombok.AllArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 
 @Service
 @AllArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE,makeFinal = true)
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class LikeService implements ILikeService {
     LikeRepository likeRepository;
     PostRepository postRepository;
     CommentRepository commentRepository;
     UserRepository userRepository;
 
+    private User getAuthenticatedUser(Authentication authentication) {
+        return userRepository.findByEmail(authentication.getName())
+                .orElseThrow(() -> new CustomException("User not found", HttpStatus.NOT_FOUND));
+    }
+
+    private Like checkDuplicateLike(long userId, Long postId, Long commentId) {
+        if (postId != null) {
+            return likeRepository.findByUserIdAndPostId(userId, postId);
+        } else {
+            return likeRepository.findByUserIdAndCommentId(userId, commentId);
+        }
+    }
+
     @Override
-    public ResponseEntity<?> likePost(long userId, long postId) {
+    public ResponseEntity<?> likePost(Authentication authentication, long postId) {
         Post post = postRepository.findById(postId)
-                .orElseThrow(()->new CustomException("The post in not found", HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new CustomException("The post is not found", HttpStatus.NOT_FOUND));
 
-        User user = userRepository.findById(userId).get();
+        User user = getAuthenticatedUser(authentication);
 
-        Like findLike = likeRepository.findByUserIdAndPostId(userId, postId);
-        if(findLike != null) {
-            throw new CustomException("Like is duplicate!",HttpStatus.BAD_REQUEST);
+        if ("Private".equals(post.getStatus()) && user != post.getUser()) {
+            throw new CustomException("Post is private, cannot like!", HttpStatus.CONFLICT);
+        }
+
+        if (checkDuplicateLike(user.getId(), postId, null) != null) {
+            throw new CustomException("Like is duplicate!", HttpStatus.BAD_REQUEST);
         }
 
         Like like = Like.builder()
@@ -43,22 +61,20 @@ public class LikeService implements ILikeService {
                 .user(user)
                 .build();
 
-        post.setLikeCount(post.getLikeCount()+1);
-
+        post.setLikeCount(post.getLikeCount() + 1);
         likeRepository.save(like);
         return ResponseEntity.ok(new Notice("Like post is completed"));
     }
 
     @Override
-    public ResponseEntity<?> likeComment(long userId, long commentId) {
+    public ResponseEntity<?> likeComment(Authentication authentication, long commentId) {
         Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(()->new CustomException("The comment in not found", HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new CustomException("The comment is not found", HttpStatus.NOT_FOUND));
 
-        User user = userRepository.findById(userId).get();
+        User user = getAuthenticatedUser(authentication);
 
-        Like findLike = likeRepository.findByUserIdAndCommentId(userId, commentId);
-        if(findLike != null) {
-            throw new CustomException("Like is duplicate!",HttpStatus.BAD_REQUEST);
+        if (checkDuplicateLike(user.getId(), null, commentId) != null) {
+            throw new CustomException("Like is duplicate!", HttpStatus.BAD_REQUEST);
         }
 
         Like like = Like.builder()
@@ -67,26 +83,32 @@ public class LikeService implements ILikeService {
                 .user(user)
                 .build();
 
-        comment.setLikeCount(comment.getLikeCount()+1);
-
+        comment.setLikeCount(comment.getLikeCount() + 1);
         likeRepository.save(like);
         return ResponseEntity.ok(new Notice("Like comment is completed"));
     }
 
     @Override
-    public ResponseEntity<?> unlike(long likeId) {
-        Like like = likeRepository.findById(likeId).get();
+    public ResponseEntity<?> unlike(Authentication authentication, long likeId) {
+        User user = getAuthenticatedUser(authentication);
+
+        Like like = likeRepository.findById(likeId)
+                .orElseThrow(() -> new CustomException("Like not found", HttpStatus.NOT_FOUND));
+
+        if (user != like.getUser()) {
+            throw new AccessDeniedException("You do not have access");
+        }
 
         Post post = like.getPost();
-
-        if(post!=null){
-            post.setLikeCount(post.getLikeCount()-1);
-        }else{
+        if (post != null) {
+            post.setLikeCount(post.getLikeCount() - 1);
+        } else {
             Comment comment = like.getComment();
-            comment.setLikeCount(comment.getLikeCount()-1);
+            comment.setLikeCount(comment.getLikeCount() - 1);
         }
 
         likeRepository.delete(like);
         return ResponseEntity.ok(new Notice("Delete like is completed"));
     }
 }
+
