+++ demo/src/main/java/com/hoanghaidang/social_network/service/impl/PostService.java
@@ -13,23 +13,37 @@ import lombok.AllArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 
 @Service
 @AllArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE,makeFinal = true)
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class PostService implements IPostService {
     PostRepository postRepository;
     UserRepository userRepository;
 
+    private User getAuthenticatedUser(Authentication authentication) {
+        return userRepository.findByEmail(authentication.getName())
+                .orElseThrow(() -> new CustomException("User not found", HttpStatus.NOT_FOUND));
+    }
+
+    private void checkPostOwnership(User user, Post post) {
+        if (post.getUser() != user) {
+            throw new AccessDeniedException("You do not have access");
+        }
+    }
+
     @Override
-    public ResponseEntity<?> createPost(PostDto postDto) {
-        if(postDto.getContent()==null && postDto.getTitle() == null && postDto.getImage() == null){
+    public ResponseEntity<?> createPost(Authentication authentication, PostDto postDto) {
+        if (postDto.getContent() == null && postDto.getTitle() == null && postDto.getImage() == null) {
             throw new CustomException("Post is required a content or a title or images", HttpStatus.BAD_REQUEST);
         }
-        User user = userRepository.findUserById(postDto.getUserId()).get();
+
+        User user = getAuthenticatedUser(authentication);
 
         Post post = Post.builder()
                 .title(postDto.getTitle())
@@ -45,23 +59,33 @@ public class PostService implements IPostService {
     }
 
     @Override
-    public ResponseEntity<?> editPost(long postId,PostDto postDto) {
-        Post post = postRepository.findById(postId).get();
+    public ResponseEntity<?> editPost(Authentication authentication, long postId, PostDto postDto) {
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new CustomException("Post could not be found", HttpStatus.NOT_FOUND));
+
+        User user = getAuthenticatedUser(authentication);
+        checkPostOwnership(user, post);
 
         post.setTitle(postDto.getTitle());
         post.setContent(postDto.getContent());
         post.setStatus(postDto.getStatus());
         post.setUpdateAt(LocalDateTime.now());
         post.setImage(postDto.getImage());
+
         postRepository.save(post);
         return ResponseEntity.ok(postDto);
     }
 
     @Override
-    public ResponseEntity<?> deletePost(long postId) {
-        Post post = postRepository.findById(postId).get();
+    public ResponseEntity<?> deletePost(Authentication authentication, long postId) {
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new CustomException("Post could not be found", HttpStatus.NOT_FOUND));
+
+        User user = getAuthenticatedUser(authentication);
+        checkPostOwnership(user, post);
 
         postRepository.delete(post);
         return ResponseEntity.ok(new Notice("Delete post completed"));
     }
 }
+
